fix: PR changes and type fix
